191111| 로그인, 회원가입, 탈퇴

+++ 05_Django/05_django_form/articles/views.py
@@ -1,4 +1,5 @@
 from django.views.decorators.http import require_POST
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from IPython import embed
 from .models import Article, Comment
@@ -6,10 +7,13 @@ from .forms import ArticleForm, CommentForm
 
 # Create your views here.
 def index(request):
+    # embed()
+    # request.session._session 을 치면 session을 볼 수 있다.
     articles = Article.objects.all()[::-1]
     context = {'articles' : articles}
     return render(request, 'articles/index.html', context)
 
+@login_required
 def create(request):
     # POST 요청 -> 데이터를 받아서 DB에 저장
     if request.method == 'POST':
@@ -46,10 +50,12 @@ def detail(request, article_pk):
 
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
     return redirect('articles:index')
 
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'POST':
@@ -80,18 +86,20 @@ def update(request, article_pk):
 @require_POST
 def comments_create(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        # save 메서드 -> 선택 인자 : (기본값) commit=True
-        # DB에 바로 저장되는 것을 막아준다.
-        comment = comment_form.save(commit=False)
-        comment.article = article
-        comment.save()
+    if request.user.is_authenticated:
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            # save 메서드 -> 선택 인자 : (기본값) commit=True
+            # DB에 바로 저장되는 것을 막아준다.
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.save()
     return redirect('articles:detail', article.pk)
     
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
     return redirect('articles:detail', article_pk)
